backend/pyproject.toml
[project]
name = "bedflow"
version = "0.1.0"
requires-python = ">=3.12,<3.15"
dependencies = [
    "celery>=5.3.6",
    "channels==4.3.1",
    "channels-redis==4.3.0",
    "cryptography>=46.0.5",
    "daphne==4.2.1",
    "django>=4.2,<5.0",
    "django-activity-stream>=1.4.0",
    "django-celery-beat>=2.5.0",
    "django-cities-light==3.10.2",
    "django-cors-headers>=4.3.1",
    "django-countries==7.6.1",
    "django-countries-plus==2.2.0",
    "django-csp>=4.0",
    "django-dotenv==1.4.2",
    "django-extensions==4.1",
    "django-filter>=23.2",
    "django-health-check>=3.17.0",
    "django-inlinecss",
    "django-jet==1.0.8",
    "django-money==3.5.4",
    "django-phonenumber-field>=8.0,<=8.5",
    "django-redis==6.0.0",
    "django-rest-passwordreset>=1.3.0",
    "django-storages>=1.14.6",
    "django-summernote>=0.8.20.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt==5.5.0",
    "drf-spectacular[sidecar]==0.28.0",
    "easy-thumbnails>=2.8.5",
    "google-auth==2.41.1",
    "gspread==6.2.1",
    "gspread-formatting==1.2.1",
    "gunicorn==23.0.0",
    "ipython==9.4",
    "paystack-sdk==1.0.1",
    "phonenumbers>=8.13,<=9.1",
    "psycopg2-binary>=2.9",
    "pyconstring>=1.0.0,<=1.4",
    "pycryptodome>=3.23,<=3.30",
    "pyjwt>=2.9.0",
    "pyotp>=2.9.0,<=3.5",
    "qrcode>=8.1,<=8.9",
    "sentry-sdk==2.33",
    "social-auth-app-django>=5.4.0",
    "twilio>=9.6,<10.0",
    "uuid7>=0.1.0",
    "watchfiles==1.1.1",
    "webauthn>=2.7.1",
    "whitenoise==6.9.0",
]

[tool.uv.sources]
django-inlinecss = { git = "https://github.com/roverdotcom/django-inlinecss.git", rev = "master" }
